Update amazon-web-service-tutorial.md
Step 8 of the Entra ID SAML SSO mentions roles. This is causing customers to be confused thinking they need to set up the roles in Entra ID. Roles are setup in AWS later in the document, then provisioned in the provisioning step. This step needs to be adjusted to reduce the confusion on the customer side regarding roles